+++ app/(security)/neuvisLanding.tsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import { Image, StyleSheet, View, Text, ScrollView, TouchableOpacity, ActivityIndicator } from 'react-native';
+import { Image, StyleSheet, View, Text, ScrollView, TouchableOpacity, 
+         ActivityIndicator, Modal, Pressable } from 'react-native';
 import Header from '../../components/Header';
 import { useNavigation, useRouter } from 'expo-router';
 import { Ionicons } from '@expo/vector-icons';
@@ -9,6 +10,18 @@ import Footer from '@/components/Footer';
 import { visit } from './types/visits';
 import { fetchExpiredUntimedoutVisitsWithNoNotificationsSentYet } from './api/notification-service/visits';
 import { insertVisitExpirationNotificationWithoutTimeout, fetchUserNotifications } from './api/notification-service/notification';
+import NotificationPanel from '../../components/NotificationPanel';
+import NotificationAlert from '../../components/NotificationAlert';
+
+//Notification
+interface Notification {
+  id: string;
+  content: string;
+  read: boolean;
+  created_at: string;
+  user_id: string;
+  visit_id?: string;
+}
 
 export default function NeuvisLanding() {
   const [expiredUntimedoutVisitswithoutNotificationsSentYet, setExpiredUntimedoutVisitsWithoutNotificationsSentYet] = useState<visit[]>([]);
@@ -20,6 +33,12 @@ export default function NeuvisLanding() {
     totalCount: 0,
     loading: true
   });
+  
+  // Notification states
+  const [notificationModalVisible, setNotificationModalVisible] = useState(false);
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+  const [showNotificationAlert, setShowNotificationAlert] = useState(false);
+  const [latestNotification, setLatestNotification] = useState<Notification | null>(null);
 
   useEffect(() => {
     navigation.setOptions({ headerShown: false });
@@ -66,6 +85,142 @@ export default function NeuvisLanding() {
     }
   };
 
+  // Fetch initial notifications and set up real-time subscription
+  useEffect(() => {
+    const fetchInitialNotifications = async () => {
+      if (user && user.id) {
+        try {
+          const { data, error } = await supabase
+            .from('notifications')
+            .select('*')
+            .eq('user_id', user.id)
+            .eq('read', false)
+            .order('created_at', { ascending: false });
+          
+          if (error) {
+            console.error('Error fetching notifications:', error);
+            return;
+          }
+          
+          if (data) {
+            setNotifications(data as Notification[]);
+            
+            // Show the most recent notification as an alert if it exists
+            if (data.length > 0) {
+              showNewNotificationAlert(data[0] as Notification);
+            }
+          }
+        } catch (error) {
+          console.error('Failed to fetch notifications:', error);
+        }
+      }
+    };
+    
+    fetchInitialNotifications();
+    
+    // Set up real-time subscription for new notifications
+    const setupNotificationSubscription = () => {
+      if (!user?.id) return null;
+      
+      return supabase
+        .channel('notifications_changes')
+        .on('postgres_changes', {
+          event: 'INSERT',
+          schema: 'public',
+          table: 'notifications',
+          filter: `user_id=eq.${user.id}`,
+        }, (payload) => {
+          // Handle new notification directly
+          const newNotification = payload.new as unknown as Notification;
+          console.log("New notification received:", newNotification);
+          
+          // Add to notifications array
+          setNotifications(prev => [newNotification, ...prev]);
+          
+          // Show notification alert - this is critical for the popup!
+          showNewNotificationAlert(newNotification);
+        })
+        .on('postgres_changes', {
+          event: 'UPDATE',
+          schema: 'public',
+          table: 'notifications',
+          filter: `user_id=eq.${user.id}`,
+        }, (payload) => {
+          // Handle updated notification directly
+          const updatedNotification = payload.new as unknown as Notification;
+          
+          // Update the notification in our local state
+          setNotifications(prev => 
+            prev.map(notification => 
+              notification.id === updatedNotification.id ? updatedNotification : notification
+            ).filter(notification => !notification.read)
+          );
+        })
+        .on('postgres_changes', {
+          event: 'DELETE',
+          schema: 'public',
+          table: 'notifications',
+          filter: `user_id=eq.${user.id}`,
+        }, (payload) => {
+          // Handle deleted notification
+          const deletedNotification = payload.old as unknown as Notification;
+          
+          // Remove from our local state
+          setNotifications(prev => 
+            prev.filter(notification => notification.id !== deletedNotification.id)
+          );
+        })
+        .subscribe();
+    };
+    
+    const subscription = setupNotificationSubscription();
+    
+    return () => {
+      if (subscription) {
+        supabase.removeChannel(subscription);
+      }
+    };
+  }, [user]);
+
+  // Function to show notification alert for new notifications
+  const showNewNotificationAlert = (notification: Notification) => {
+    console.log("Showing notification alert for:", notification.content);
+    
+    // Only show alert for unread notifications
+    if (!notification.read) {
+      setLatestNotification(notification);
+      setShowNotificationAlert(true);
+    }
+  };
+
+  // Handle notification icon press
+  const handleNotificationPress = () => {
+    setNotificationModalVisible(prev => !prev); // Toggle notification panel
+  };
+
+  // For testing - show a test notification alert
+  const testNotificationAlert = () => {
+    const testNotification: Notification = {
+      id: "test-" + Date.now().toString(),  // Generate unique string ID
+      content: "This is a test notification to verify the alert is working correctly.",
+      read: false,
+      created_at: new Date().toISOString(),
+      user_id: user?.id || "test-user"
+    };
+    
+    console.log("Triggering test notification alert");
+    // Add to notifications array so it appears in the list too
+    setNotifications(prev => [testNotification, ...prev]);
+    showNewNotificationAlert(testNotification);
+  };
+
+  // Handle when notification alert is closed
+  const handleHideNotificationAlert = () => {
+    console.log("Hiding notification alert");
+    setShowNotificationAlert(false);
+    setLatestNotification(null);
+  };
+
   useEffect(() => {
     const sendNotfications = async () => {
       const data = await fetchExpiredUntimedoutVisitsWithNoNotificationsSentYet(user);
@@ -85,6 +240,7 @@ export default function NeuvisLanding() {
     sendNotfications();
   }, []);
 
+  const unreadCount = notifications.filter(n => !n.read).length;
   
   return (
     <View style={styles.mainContainer}>
@@ -98,7 +254,7 @@ export default function NeuvisLanding() {
           <Text style={styles.heroSubtitle}>Scan a valid ID or manually enter visitor details.</Text>
         </View>
 
-        {/* Simple Stats Section */}
+        {/* Simple Stats Section - Notification button moved to right controls */}
         <View style={styles.statsWrapper}>
           {stats.loading ? (
             <View style={styles.loadingContainer}>
@@ -115,12 +271,29 @@ export default function NeuvisLanding() {
                 <Text style={styles.statNumber}>{stats.totalCount}</Text>
                 <Text style={styles.statLabel}>Total Visitors</Text>
               </View>
-              <TouchableOpacity 
-                style={styles.refreshButton}
-                onPress={fetchSimpleStats}
-              >
-                <Ionicons name="refresh-outline" size={16} color="#003566" />
-              </TouchableOpacity>
+              <View style={styles.rightControlsContainer}>
+                <TouchableOpacity 
+                  style={styles.refreshButton}
+                  onPress={fetchSimpleStats}
+                >
+                  <Ionicons name="refresh-outline" size={16} color="#003566" />
+                </TouchableOpacity>
+                {/* Notification bell moved here */}
+                <TouchableOpacity 
+                  style={styles.notificationButton}
+                  onPress={handleNotificationPress}
+                  onLongPress={testNotificationAlert} // Long press to test notification
+                >
+                  <Ionicons name="notifications" size={16} color="#003566" />
+                  {unreadCount > 0 && (
+                    <View style={styles.notificationBadge}>
+                      <Text style={styles.notificationCount}>
+                        {unreadCount > 99 ? '99+' : unreadCount}
+                      </Text>
+                    </View>
+                  )}
+                </TouchableOpacity>
+              </View>
             </View>
           )}
         </View>
@@ -189,6 +362,24 @@ export default function NeuvisLanding() {
         </View>
         <Footer />
       </ScrollView>
+
+      {/* Notification Panel */}
+      <NotificationPanel
+        visible={notificationModalVisible}
+        onClose={() => setNotificationModalVisible(false)}
+        user={user}
+        notifications={notifications}
+        setNotifications={setNotifications}
+      />
+
+      {/* Notification Alert - Make sure it's the last component to ensure proper z-index layering */}
+      {showNotificationAlert && latestNotification && (
+        <NotificationAlert
+          notification={latestNotification}
+          visible={showNotificationAlert}
+          onHide={handleHideNotificationAlert}
+        />
+      )}
     </View>
   );
 }
@@ -259,22 +450,54 @@ const styles = StyleSheet.create({
     fontSize: 28,
     fontWeight: 'bold',
     color: '#252525',
+    left: -10
   },
   statLabel: {
     fontSize: 14,
     color: '#252525',
     marginTop: 4,
     opacity: 0.7,
+    
   },
   statDivider: {
-    width: 1,
+    width: 2,
     height: 50,
     backgroundColor: '#eeeeee',
   },
-  refreshButton: {
+  rightControlsContainer: {
     position: 'absolute',
     right: 0,
     top: 0,
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  notificationButton: {
+    position: 'relative',
+    padding: 6,
+    marginLeft: 1,
+    left: 5
+  },
+  notificationBadge: {
+    position: 'absolute',
+    top: -4,
+    right: -4,
+    backgroundColor: '#ff3b30',
+    borderRadius: 10,
+    minWidth: 20,
+    height: 20,
+    justifyContent: 'center',
+    alignItems: 'center',
+    paddingHorizontal: 4,
+    borderWidth: 1.5,
+    borderColor: '#fff', // Add white border for better contrast
+  },
+  notificationCount: {
+    color: '#fff',
+    fontSize: 10,
+    fontWeight: 'bold',
+    textAlign: 'center', // Ensure text is centered
+  },
+  refreshButton: {
     padding: 6,
     backgroundColor: '#f8f9fa',
     borderRadius: 20,
