+++ components/Header.tsx
@@ -6,35 +6,17 @@ import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { router } from 'expo-router';
 import { useAuth } from '@/app/context/AuthContext';
 import { supabase } from '@/app/lib/supabaseClient';
-import NotificationPanel from './NotificationPanel';
-import NotificationAlert from './NotificationAlert';
-
-// Define the Notification interface
-interface Notification {
-  id: string;  // Using string type for id
-  content: string;
-  read: boolean;
-  created_at: string;
-  user_id: string;
-  visit_id?: string;
-}
-
+  
 interface HeaderProps {
   onProfilePress?: () => void;
-  onNotificationPress?: () => void;
 }
 
 const Header: React.FC<HeaderProps> = ({ 
   onProfilePress,
-  onNotificationPress,
 }) => {
   const { user, signOut } = useAuth();
   const insets = useSafeAreaInsets();
   const [profileModalVisible, setProfileModalVisible] = useState(false);
-  const [notificationModalVisible, setNotificationModalVisible] = useState(false);
-  const [notifications, setNotifications] = useState<Notification[]>([]);
-  const [showNotificationAlert, setShowNotificationAlert] = useState(false);
-  const [latestNotification, setLatestNotification] = useState<Notification | null>(null);
   const [userData, setUserData] = useState({
     full_name: "Loading...",
     role: "Loading..."
@@ -71,147 +53,6 @@ const Header: React.FC<HeaderProps> = ({
     fetchUserData();
   }, [user]);
 
-  // Fetch initial notifications and set up real-time subscription
-  useEffect(() => {
-    const fetchInitialNotifications = async () => {
-      if (user && user.id) {
-        try {
-          const { data, error } = await supabase
-            .from('notifications')
-            .select('*')
-            .eq('user_id', user.id)
-            .eq('read', false)
-            .order('created_at', { ascending: false });
-          
-          if (error) {
-            console.error('Error fetching notifications:', error);
-            return;
-          }
-          
-          if (data) {
-            setNotifications(data as Notification[]);
-            
-            // Show the most recent notification as an alert if it exists
-            if (data.length > 0) {
-              showNewNotificationAlert(data[0] as Notification);
-            }
-          }
-        } catch (error) {
-          console.error('Failed to fetch notifications:', error);
-        }
-      }
-    };
-    
-    fetchInitialNotifications();
-    
-    // Set up real-time subscription for new notifications
-    const setupNotificationSubscription = () => {
-      if (!user?.id) return null;
-      
-      return supabase
-        .channel('notifications_changes')
-        .on('postgres_changes', {
-          event: 'INSERT',
-          schema: 'public',
-          table: 'notifications',
-          filter: `user_id=eq.${user.id}`,
-        }, (payload) => {
-          // Handle new notification directly
-          const newNotification = payload.new as unknown as Notification;
-          console.log("New notification received:", newNotification);
-          
-          // Add to notifications array
-          setNotifications(prev => [newNotification, ...prev]);
-          
-          // Show notification alert - this is critical for the popup!
-          showNewNotificationAlert(newNotification);
-        })
-        .on('postgres_changes', {
-          event: 'UPDATE',
-          schema: 'public',
-          table: 'notifications',
-          filter: `user_id=eq.${user.id}`,
-        }, (payload) => {
-          // Handle updated notification directly
-          const updatedNotification = payload.new as unknown as Notification;
-          
-          // Update the notification in our local state
-          setNotifications(prev => 
-            prev.map(notification => 
-              notification.id === updatedNotification.id ? updatedNotification : notification
-            ).filter(notification => !notification.read)
-          );
-        })
-        .on('postgres_changes', {
-          event: 'DELETE',
-          schema: 'public',
-          table: 'notifications',
-          filter: `user_id=eq.${user.id}`,
-        }, (payload) => {
-          // Handle deleted notification
-          const deletedNotification = payload.old as unknown as Notification;
-          
-          // Remove from our local state
-          setNotifications(prev => 
-            prev.filter(notification => notification.id !== deletedNotification.id)
-          );
-        })
-        .subscribe();
-    };
-    
-    const subscription = setupNotificationSubscription();
-    
-    return () => {
-      if (subscription) {
-        supabase.removeChannel(subscription);
-      }
-    };
-  }, [user]);
-
-  // Function to show notification alert for new notifications
-  const showNewNotificationAlert = (notification: Notification) => {
-    console.log("Showing notification alert for:", notification.content);
-    
-    // Only show alert for unread notifications
-    if (!notification.read) {
-      setLatestNotification(notification);
-      setShowNotificationAlert(true);
-    }
-  };
-
-  // Handle notification icon press
-  const handleNotificationPress = () => {
-    setNotificationModalVisible(prev => !prev); // Toggle notification panel
-    if (onNotificationPress) {
-      onNotificationPress();
-    }
-  };
-
-  // For testing - show a test notification alert
-  const testNotificationAlert = () => {
-    const testNotification: Notification = {
-      id: "test-" + Date.now().toString(),  // Generate unique string ID
-      content: "This is a test notification to verify the alert is working correctly.",
-      read: false,
-      created_at: new Date().toISOString(),
-      user_id: user?.id || "test-user"
-    };
-    
-    console.log("Triggering test notification alert");
-    // Add to notifications array so it appears in the list too
-    setNotifications(prev => [testNotification, ...prev]);
-    showNewNotificationAlert(testNotification);
-  };
-
-  // Handle when notification alert is closed
-  const handleHideNotificationAlert = () => {
-    console.log("Hiding notification alert");
-    setShowNotificationAlert(false);
-    setLatestNotification(null);
-  };
-
-  const unreadCount = notifications.filter(n => !n.read).length;
-
   const handleViewProfile = () => {
     setProfileModalVisible(false);
     // Navigate to ProfileEditor
@@ -231,21 +72,6 @@ const Header: React.FC<HeaderProps> = ({
       </View>
 
       <View style={styles.right}>
-        <TouchableOpacity 
-          style={styles.notificationButton}
-          onPress={handleNotificationPress}
-          onLongPress={testNotificationAlert} // Long press to test notification
-        >
-          <Ionicons name="notifications" size={20} color="#252525" />
-          {unreadCount > 0 && (
-            <View style={styles.notificationBadge}>
-              <Text style={styles.notificationCount}>
-                {unreadCount > 99 ? '99+' : unreadCount}
-              </Text>
-            </View>
-          )}
-        </TouchableOpacity>
-
         <TouchableOpacity 
           style={styles.profileButton}
           onPress={() => setProfileModalVisible(true)}
@@ -312,24 +138,6 @@ const Header: React.FC<HeaderProps> = ({
           </View>
         </Pressable>
       </Modal>
-
-      {/* Notification Panel */}
-      <NotificationPanel
-        visible={notificationModalVisible}
-        onClose={() => setNotificationModalVisible(false)}
-        user={user}
-        notifications={notifications}
-        setNotifications={setNotifications}
-      />
-
-      {/* Notification Alert - Make sure it's the last component to ensure proper z-index layering */}
-      {showNotificationAlert && latestNotification && (
-        <NotificationAlert
-          notification={latestNotification}
-          visible={showNotificationAlert}
-          onHide={handleHideNotificationAlert}
-        />
-      )}
     </View>
   );
 };
@@ -368,33 +176,6 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
     alignItems: 'center',
   },
-  notificationButton: {
-    position: 'relative',
-    marginRight: 12,
-    padding: 8,
-    zIndex: 11, 
-  },
-  notificationBadge: {
-    position: 'absolute',
-    top: -2,
-    right: -2,
-    backgroundColor: '#ff3b30',
-    borderRadius: 10,
-    minWidth: 20,
-    height: 20,
-    justifyContent: 'center',
-    alignItems: 'center',
-    paddingHorizontal: 4,
-    borderWidth: 1.5,
-    borderColor: '#fff', // Add white border for better contrast
-  },
-  notificationCount: {
-    color: '#fff',
-    fontSize: 10,
-    fontWeight: 'bold',
-    textAlign: 'center', // Ensure text is centered
-  },
-  
   profileButton: {
     flexDirection: 'row',
     alignItems: 'center',
@@ -402,7 +183,7 @@ const styles = StyleSheet.create({
     borderRadius: 20,
     paddingVertical: 5,
     paddingHorizontal: 8,
-  },
+  },// Define the Notification interface
   avatar: {
     width: 28,
     height: 28,
